+++ BitcoinMillionaire/BitcoinMillionaire/Utilities/Constants.swift
@@ -13,7 +13,7 @@ struct Constants {
         static let bitcoinPriceEndpoint = "https://api.coindesk.com/v1/bpi/currentprice.json"
     }
     
-    static let oneMillionDollars: Float = 1_000_000
+    static let oneMillionDollars: Double = 1_000_000
     
 }
